LayoutTests/fast/dom/Geolocation/permission-denied-already-error-expected.txt

Tests that when Geolocation permission has been denied prior to a call to a Geolocation method, the error callback is invoked with code PERMISSION_DENIED, when the Geolocation service encounters an error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true
PASS error.code is error.PERMISSION_DENIED
PASS error.message is "User denied Geolocation"
PASS error.UNKNOWN_ERROR is 0
PASS error.PERMISSION_DENIED is 1
PASS error.POSITION_UNAVAILABLE is 2
PASS error.TIMEOUT is 3

PASS error.code is error.PERMISSION_DENIED
PASS error.message is "User denied Geolocation"
PASS error.UNKNOWN_ERROR is 0
PASS error.PERMISSION_DENIED is 1
PASS error.POSITION_UNAVAILABLE is 2
PASS error.TIMEOUT is 3

TEST COMPLETE
